requirements.txt
asgiref==3.11.0
Django==6.0.1
django-axes==8.1.0
django-cors-headers==4.9.0
django-rest-knox==5.0.2
djangorestframework==3.16.1
pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
